Replace ServiceCredentialsRef with ServiceUser
The problem we have to create the secrets inside the ns of the deployment they
belong to is that we dont know the namespace until we assign a ServiceUser into
a ServiceCandidate. Before we were always creating the secrets fields once the
SDPUsers were created in the SDP system, this can not be done anymore since we
dont know the ns at that point.

Now it will be created when the ServiceUser is activated after it has been
assigned to a ServiceCandidate.