Allow IC to delete/create entries in the glocal secret managed by the service
We need to store somewhere the passwords fromfor the SDPUsers we create because
controller won't return its value next time we fetch them. IC is taking care of
that, it creates it when a new SDPService is derived from the SDPUser and it
deletes it when a SDPService is deleted.